requirements.txt
asgiref==3.8.1
bcrypt==4.1.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.7
Django==4.2
django-appconf==1.0.6
django-cors-headers==4.3.1
django-ratelimit==4.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
idna==3.8
openpyxl==3.1.5
pandas==2.2.2
packaging==24.1
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2023.3
razorpay==1.4.2
requests==2.31.0
setuptools==73.0.1
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1
